requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
pytz==2019.3
sqlparse==0.3.1
